requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --all-extras pyproject.toml
#
altair==5.2.0
    # via streamlit
annotated-types==0.6.0
    # via pydantic
astroid==3.1.0
    # via pylint
attrs==23.1.0
    # via
    #   jsonschema
    #   referencing
black==24.3.0
    # via dkp (pyproject.toml)
blinker==1.7.0
    # via streamlit
cachetools==5.3.2
    # via streamlit
certifi==2023.11.17
    # via requests
cffi==1.16.0
    # via
    #   cryptography
    #   pynacl
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   black
    #   streamlit
coverage[toml]==7.4.4
    # via
    #   coverage
    #   pytest-cov
cryptography==42.0.5
    # via pyjwt
deprecated==1.2.14
    # via pygithub
dill==0.3.8
    # via pylint
gitdb==4.0.11
    # via gitpython
gitpython==3.1.40
    # via streamlit
idna==3.6
    # via requests
iniconfig==2.0.0
    # via pytest
isort==5.13.2
    # via
    #   dkp (pyproject.toml)
    #   pylint
jinja2==3.1.2
    # via
    #   altair
    #   pydeck
jsonschema==4.20.0
    # via altair
jsonschema-specifications==2023.12.1
    # via jsonschema
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.3
    # via jinja2
mccabe==0.7.0
    # via pylint
mdurl==0.1.2
    # via markdown-it-py
mypy-extensions==1.0.0
    # via black
numpy==1.26.2
    # via
    #   altair
    #   pandas
    #   pyarrow
    #   pydeck
    #   streamlit
oauthlib==3.2.2
    # via requests-oauthlib
packaging==23.2
    # via
    #   altair
    #   black
    #   pytest
    #   streamlit
pandas==2.1.4
    # via
    #   altair
    #   streamlit
pathspec==0.12.1
    # via black
pillow==10.1.0
    # via streamlit
platformdirs==4.2.0
    # via
    #   black
    #   pylint
pluggy==1.4.0
    # via pytest
protobuf==4.25.1
    # via streamlit
pyarrow==14.0.2
    # via streamlit
pycparser==2.21
    # via cffi
pydantic==2.6.4
    # via dkp (pyproject.toml)
pydantic-core==2.16.3
    # via pydantic
pydeck==0.8.1b0
    # via streamlit
pygithub==2.3.0
    # via dkp (pyproject.toml)
pygments==2.17.2
    # via rich
pyjwt[crypto]==2.8.0
    # via
    #   pygithub
    #   pyjwt
pylint==3.1.0
    # via dkp (pyproject.toml)
pynacl==1.5.0
    # via pygithub
pytest==8.1.1
    # via
    #   dkp (pyproject.toml)
    #   pytest-cov
    #   pytest-mock
pytest-cov==5.0.0
    # via dkp (pyproject.toml)
pytest-mock==3.14.0
    # via dkp (pyproject.toml)
python-dateutil==2.8.2
    # via pandas
python-dotenv==1.0.1
    # via dkp (pyproject.toml)
pytz==2023.3.post1
    # via pandas
pyyaml==6.0.1
    # via dkp (pyproject.toml)
referencing==0.32.0
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.31.0
    # via
    #   dkp (pyproject.toml)
    #   pygithub
    #   requests-oauthlib
    #   streamlit
requests-oauthlib==2.0.0
    # via dkp (pyproject.toml)
rich==13.7.0
    # via streamlit
rpds-py==0.15.2
    # via
    #   jsonschema
    #   referencing
six==1.16.0
    # via python-dateutil
smmap==5.0.1
    # via gitdb
streamlit==1.32.2
    # via dkp (pyproject.toml)
tenacity==8.2.3
    # via streamlit
toml==0.10.2
    # via streamlit
tomlkit==0.12.4
    # via pylint
toolz==0.12.0
    # via altair
tornado==6.4
    # via streamlit
typing-extensions==4.9.0
    # via
    #   pydantic
    #   pydantic-core
    #   pygithub
    #   streamlit
tzdata==2023.3
    # via pandas
urllib3==2.0.7
    # via
    #   pygithub
    #   requests
watchdog==3.0.0
    # via streamlit
wrapt==1.16.0
    # via deprecated
